feat(cs)/implement user banning functionality

+++ services/core-service/src/domain/services/user-management.service.ts
@@ -7,27 +7,35 @@ export class UserManagementService {
 	private userRepository = new UserRepository();
 	private authSessionRepository = new AuthSessionRepository();
 
-	public getUsers() {
-		return this.userRepository.getUsers();
+	public async getUsers() {
+		return (await this.userRepository.getUsers()).filter(u => u.role == 'USER');
 	}
 
 	public async removeUser(userId: string) {
 		const user = await this.userRepository.getUserById(userId);
 		if (!user) throw new ApiError(404, 'User does not exist');
+		if (user.role == 'ADMIN') throw new ApiError(403, 'Administrators cannot be removed');
 
 		await this.userRepository.deleteUserById(userId);
 
 		return user;
 	}
 
-	public banUser(userId: string) {
-		//TODO: Set user banned
-		//TODO: Close user sessions
+	public async banUser(userId: string) {
+		const user = await this.userRepository.getUserById(userId);
+		if (!user) throw new ApiError(404, 'User does not exist');
+		if (user.role == 'ADMIN') throw new ApiError(403, 'Administrators cannot be banned');
+
+		const bannedUser = await this.userRepository.updateUserById(userId, { banned: true });
+		await this.closeSessions(userId);
+
+		return bannedUser;
 	}
 
 	public async closeSessions(userId: string) {
 		const user = await this.userRepository.getUserById(userId);
 		if (!user) throw new ApiError(404, 'User does not exist');
+		if (user.role == 'ADMIN') throw new ApiError(403, 'Administrator sessions cannot be closed');
 
 		return this.authSessionRepository.clearUserSessions(userId);
 	}
@@ -35,6 +43,7 @@ export class UserManagementService {
 	public async updateUser(userId: string, updates: Partial<user>) {
 		const user = await this.userRepository.getUserById(userId);
 		if (!user) throw new ApiError(404, 'User does not exist');
+		if (user.role == 'ADMIN') throw new ApiError(403, 'Administrators cannot be updated by other administrators');
 
 		const updated = await this.userRepository.updateUserById(userId, updates);
